+++ src/main/java/module-info.java
@@ -13,8 +13,8 @@ module org.example {
     exports org.example.service to com.fasterxml.jackson.databind;
     exports org.example.infrastructure to com.fasterxml.jackson.databind;
 
-    opens org.example.ui to javafx.fxml;
-    exports org.example.ui to javafx.graphics;
+    opens org.example.app to javafx.fxml;
+    exports org.example.app to javafx.graphics;
     exports org.example.controller to javafx.graphics;
     opens org.example.controller to javafx.fxml;
     exports org.example.manager to javafx.graphics;
