+++ src/main/java/org/example/app/TimerMode.java
@@ -1,4 +1,4 @@
-package org.example.ui;
+package org.example.app;
 
 public enum TimerMode {
     FOCUS,
