+++ src/main/java/org/example/manager/TimerManager.java
@@ -4,7 +4,7 @@ import javafx.animation.KeyFrame;
 import javafx.animation.Timeline;
 import javafx.util.Duration;
 import org.example.model.dto.TimerDetails;
-import org.example.ui.TimerMode;
+import org.example.app.TimerMode;
 
 public class TimerManager {
